integrations/google-adk/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cognee-integration-google-adk"
version = "0.1.1"
description = "Integrate Cognee's memory system seamlessly with your Google ADK agent using easy-to-use tools."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cognee>=0.4.0,<0.5.4",
    "google-adk>=1.19.0",
    "python-dotenv>=1.2.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["cognee_integration_google_adk*"]
exclude = ["examples*"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.14.6",
]
